requirements.txt
aioredis==1.3.1
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.3.1
astor==0.8.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
autoflake==1.4
Automat==20.2.0
Babel==2.9.0
bandit==1.6.2
black==20.8b1
certifi==2020.11.8
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
coverage==5.3
cryptography==3.2.1
daphne==3.0.1
darglint==1.5.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
django-pgcrypto==2.0.0
doc8==0.8.1
docutils==0.16
eradicate==2.0.0
filelock==3.0.12
flake8==3.8.4
flake8-2020==1.6.0
flake8-bandit==2.1.2
flake8-broken-line==0.3.0
flake8-bugbear==20.1.4
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.3.0
flake8-debugger==3.2.1
flake8-docstrings==1.5.0
flake8-eradicate==1.0.0
flake8-fixme==1.1.1
flake8-isort==4.0.0
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==3.2.0
flake8-rst-docstrings==0.0.14
flake8-string-format==0.3.0
flake8-variables-names==0.0.3
gitdb==4.0.5
GitPython==3.1.11
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
imagesize==1.2.0
incremental==17.5.0
iniconfig==1.1.1
isort==5.6.4
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.1
pbr==5.5.1
pep8-naming==0.11.1
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.2
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
python-dev-tools==2020.9.10
python-dotenv==0.15.0
pytz==2020.4
pyupgrade==2.7.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
restructuredtext-lint==1.3.1
service-identity==18.1.0
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
stevedore==3.2.2
testfixtures==6.15.0
tokenize-rt==4.0.0
toml==0.10.2
tox==3.20.1
tox-travis==0.12
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.1.0
wemake-python-styleguide==0.14.1
whitenoise==5.2.0
zope.interface==5.2.0
